Excercise Completed

+++ src/main/java/org/lessons/springlamiapizzeriacrud/controller/IngridientController.java
@@ -2,12 +2,15 @@ package org.lessons.springlamiapizzeriacrud.controller;
 
 import jakarta.validation.Valid;
 import org.lessons.springlamiapizzeriacrud.model.Ingridient;
+import org.lessons.springlamiapizzeriacrud.model.Pizza;
 import org.lessons.springlamiapizzeriacrud.repository.IngridientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -50,6 +53,14 @@ public class IngridientController {
 
     @PostMapping("/delete/{id}")
     public String delete(@PathVariable Integer id) {
+        Optional<Ingridient> result = ingridientRepository.findById(id);
+        if (result.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        Ingridient ingridientToDelete = result.get();
+        for (Pizza pizza : ingridientToDelete.getPizzas()) {
+            pizza.getIngridients().remove(ingridientToDelete);
+        }
         ingridientRepository.deleteById(id);
         return "redirect:/ingridients";
     }
